split up transformer tests into separate files by topic
